Mesh/mesh intersection: use the `spade` crate for triangulation
