+++ crates/parry3d-f64/Cargo.toml
@@ -52,6 +52,7 @@ serde      = { version = "1.0", optional = true, features = ["derive", "rc"]}
 num-derive = "0.3"
 indexmap   = { version = "1", features = [ "serde-1" ], optional = true }
 rustc-hash = { version = "1", optional = true }
+spade      = { version = "2" } # Make this optional?
 
 [target.'cfg(not(target_os = "cuda"))'.dependencies]
 cust       = { version = "0.3", optional = true }
