+++ src/shape/trimesh.rs
@@ -81,6 +81,20 @@ pub struct TriMeshTopology {
     pub half_edges: Vec<TopoHalfEdge>,
 }
 
+impl TriMeshTopology {
+    pub(crate) fn face_half_edges_ids(&self, fid: u32) -> [u32; 3] {
+        let first_half_edge = self.faces[fid as usize].half_edge;
+
+        let mut result = [first_half_edge; 3];
+        for k in 1..3 {
+            let half_edge = self.half_edges[result[k - 1] as usize];
+            result[k] = half_edge.next;
+        }
+
+        result
+    }
+}
+
 #[cfg(feature = "dim3")]
 impl Default for PseudoNormals {
     fn default() -> Self {
